requirements.txt
Django
django-bootstrap5
django-tables2
pandas
psycopg2-binary
djangorestframework
tzdata
django-extensions
Werkzeug
pyOpenSSL
python-barcode
pillow